Rennes_metro_project/web/requirements.txt
Flask
mysql-connector-python
Flask-Cors

